+++ src/year_2015/day_05.rs
@@ -88,23 +88,23 @@ fn string_is_nice_v2(input: &str) -> bool {
   twice_pair && repeated
 }
 
-pub fn day_05_v1(input: impl Into<String>) -> usize {
+pub fn day_05_v1(input: impl Into<String>) -> u16 {
   input
     .into()
     .lines()
     .filter(|line| string_is_nice_v1(line))
-    .count()
+    .count() as u16
 }
 
-pub fn day_05_v2(input: impl Into<String>) -> usize {
+pub fn day_05_v2(input: impl Into<String>) -> u16 {
   input
     .into()
     .lines()
     .filter(|line| string_is_nice_v2(line))
-    .count()
+    .count() as u16
 }
 
-solvable!(day_05, day_05_v1, day_05_v2, usize);
+solvable!(day_05, day_05_v1, day_05_v2, u16);
 
 #[cfg(test)]
 mod tests {
